// packages/machines/date-time-picker/src/date-time-picker.types.ts
export type DateView = "day" | "month" | "year"